Update manager_effect_35E.lua
Add O modifer to effect parsing (alongside HTQ). The O modifier acts in the same way as the H modifier but will round the value (math.ceil) rather than truncate it (math.floor). This allows an effect like Holy Vindicator - Vindicator's Shield to be coded where the bonus is equal to the number of Channel Energy dice of the character and those dice go up on odd levels.
e.g. for a character that is Cleric 9 / Holy Vindicator 4 their effective level for channel energy is 13 which gives 7d6. [HCL] would give an incorrect modifier of 6, this [OCL] change would give a correct modifier of 7.